# api/docs/swagger.yaml
basePath: /v1
definitions:
  http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
...
  models.BookReq:
    properties:
      audio_url:
        type: string
      author_id:
        type: integer
      book_type:
        example: standard
        type: string
      category_id:
        type: integer
      description:
        type: string
      download_url:
        type: string
      image:
        type: string
      sub_category_id:
        type: integer
      title:
        type: string
    type: object
  models.BookRes:
    properties:
      audio_url:
        type: string
      author_id:
        type: integer
      book_type:
        type: string
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      download_url:
        type: string
      id:
        type: integer
      image:
        type: string
      sub_category_id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.BookResponse:
    properties:
      audio_url:
        type: string
      author_first_name:
        type: string
      author_id:
        type: integer
      author_last_name:
        type: string
      book_type:
        type: string
      category_id:
        type: integer
      comment_count:
        type: integer
      comments:
        items:
          $ref: '#/definitions/models.CommentResponse'
        type: array
      created_at:
        type: string
      description:
        type: string
      download_url:
        type: string
      id:
        type: integer
      image:
        type: string
      is_like:
        type: boolean
      like_count:
        type: integer
      similar_books:
        items:
          $ref: '#/definitions/models.BooksForList'
        type: array
      sub_category_id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      view_count:
        type: integer
    type: object
  models.BookUpdate:
    properties:
      audio_url:
        type: string
      book_type:
        type: string
      description:
        type: string
      download_url:
        type: string
      id:
        type: integer
      image:
        type: string
      title:
        type: string
    type: object
  models.Books:
    properties:
      _meta:
        $ref: '#/definitions/models.Meta'
      books:
        items:
          $ref: '#/definitions/models.BooksForList'
        type: array
    type: object
  models.BooksAudioList:
    properties:
      audio_url:
        type: string
      author_first_name:
        type: string
      author_id:
        type: integer
      author_last_name:
        type: string
      book_type:
        type: string
      category_id:
        type: integer
      id:
        type: integer
      image:
        type: string
      like_count:
        type: integer
      sub_category_id:
        type: integer
      title:
        type: string
      view_count:
        type: integer
    type: object
  models.BooksAudios:
    properties:
      _meta:
        $ref: '#/definitions/models.Meta'
      books:
        items:
          $ref: '#/definitions/models.BooksAudioList'
        type: array
    type: object
  models.BooksForList:
    properties:
      author_first_name:
        type: string
      author_id:
        type: integer
      author_last_name:
        type: string
      book_type:
        type: string
      category_id:
        type: integer
      id:
        type: integer
      image:
        type: string
      like_count:
        type: integer
      sub_category_id:
        type: integer
      title:
        type: string
      view_count:
        type: integer
    type: object
...
  models.CommentResponse:
    properties:
      book_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      text:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.UserResForComment'
      user_id:
        type: integer
    type: object
...
  models.DefaultResponse:
    properties:
      error_code:
        type: integer
      error_message:
        type: string
    type: object
...
  models.LikeReq:
    properties:
      book_id:
        type: integer
      user_id:
        type: integer
    type: object
...
  models.Meta:
    properties:
      current_page:
        type: integer
      page_count:
        type: integer
      per_page:
        type: integer
      total_count:
        type: integer
    type: object
...
  models.Success:
    properties:
      message:
        type: string
    type: object
...
  models.UserResForComment:
    properties:
      avatar_url:
        type: string
      full_name:
        type: string
      id:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server.
  termsOfService: http://swagger.io/terms/
  title: Book store API
  version: "1.0"
paths:
...
  /book:
    post:
      consumes:
      - application/json
      description: Here books will be created.
      parameters:
      - description: Create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.BookReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.BookRes'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.DefaultResponse'
      security:
      - BearerAuth: []
      summary: create books
      tags:
      - Book
    put:
      consumes:
      - application/json
      description: Update Book
      operationId: updatebook
      parameters:
      - description: UpdateBook
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.BookUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Books data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.BookRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update Book
      tags:
      - Book
  /book/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Category
      operationId: deletebook
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Success'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete Category
      tags:
      - Book
    get:
      consumes:
      - application/json
      description: Here get by books.
      operationId: getbook
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: user_id
        in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Book data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.BookResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: get books
      tags:
      - Book
  /book/audios:
    get:
      consumes:
      - application/json
      description: GET audios of books
      operationId: getbookaudios
      parameters:
      - description: LIMIT
        in: query
        name: limit
        type: string
      - description: PAGE_NUMBER
        in: query
        name: page
        type: string
      - description: SEARCH
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Books data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.BooksAudios'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: GET audios
      tags:
      - Book
  /book/filter:
    get:
      consumes:
      - application/json
      description: GET Books
      operationId: getbooks
      parameters:
      - description: LIMIT
        in: query
        name: limit
        type: integer
      - description: PAGE_NUMBER
        in: query
        name: page
        type: integer
      - description: SEARCH_KEY
        in: query
        name: search
        type: string
      - description: CATEGORY_ID
        in: query
        name: category_id
        type: integer
      - description: SUBCATEGORY_ID
        in: query
        name: subcategory_id
        type: integer
      - description: AUTHOR_ID
        in: query
        name: author_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Books data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Books'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: GET Books
      tags:
      - Book
  /book/like:
    delete:
      consumes:
      - application/json
      description: Here books will be delete like.
      parameters:
      - description: DeleteLike
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.LikeReq'
      produces:
      - application/json
      responses:
        "200":
          description: DeletedLike
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.DefaultResponse'
      security:
      - BearerAuth: []
      summary: delete book like
      tags:
      - Book
    post:
      consumes:
      - application/json
      description: Here books will be created like.
      parameters:
      - description: CreateLike
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.LikeReq'
      produces:
      - application/json
      responses:
        "201":
          description: CreatedLike
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.DefaultResponse'
      security:
      - BearerAuth: []
      summary: create book like
      tags:
      - Book
  /book/mostread:
    get:
      consumes:
      - application/json
      description: GET list of most read books
      operationId: getbookmostread
      parameters:
      - description: LIMIT
        in: query
        name: limit
        type: string
      - description: PAGE_NUMBER
        in: query
        name: page
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Books data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Books'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: GET most read books
      tags:
      - Book
  /book/search:
    get:
      consumes:
      - application/json
      description: GET Books
      operationId: getbooklist
      parameters:
      - description: LIMIT
        in: query
        name: limit
        type: integer
      - description: PAGE_NUMBER
        in: query
        name: page
        type: integer
      - description: SEARCH_KEY
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Books data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Books'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: GET Books
      tags:
      - Book
  /book/top:
    get:
      consumes:
      - application/json
      description: GET TOP books
      operationId: getbooksearch
      parameters:
      - description: LIMIT
        in: query
        name: limit
        type: string
      - description: PAGE_NUMBER
        in: query
        name: page
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Books data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Books'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: GET TOP Books
      tags:
      - Book
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"